check for open sections

+++ Services/PopulationGenerating/CreatePopulation.cs
@@ -5,7 +5,7 @@ namespace Scheduler.Services.PopulationGenerating
     {
         public Dictionary<List<Section>, int> InitializePopulation(Dictionary<int, List<Section>> sectionsByCourse, int populationSize)
         {
-        HashSet<int> populationIndex = new HashSet<int>();
+        HashSet<long> populationIndex = new HashSet<long>();
         var population = new Dictionary<List<Section>,int>();
             var random = new Random();
             for (int i = 0; i < populationSize; i++)
@@ -18,9 +18,9 @@ namespace Scheduler.Services.PopulationGenerating
                     index += section.IDSection.ToString();
                     schedule.Add(section);
                 }
-                if(!populationIndex.Contains(int.Parse(index)))
+                if(!populationIndex.Contains(long.Parse(index)))
                 {
-                    populationIndex.Add(int.Parse(index));
+                    populationIndex.Add(long.Parse(index));
                     population[schedule] = 0;
                 }
             }
